Split the child-process runner out of the subagent body
Spawning a pi child, parsing its JSON event stream and tearing it down on abort
or a maxTurns cap is now one module. It is the only part of the extension that
owns a process, a temp file or a worktree.

The subagent entry point is 838 lines, from 1693 before this series.

Pure move: no behavior change.
